Bug fixed.

+++ models/models.go
@@ -7,7 +7,7 @@ type Access struct {
 
 type User struct {
 	AccessToken string
-	GithubId    int
+	GithubId    int64
 	AvatarUrl   string
 	ProfileUrl  string
 	FullName    string
